+++ .github/workflows/e2e-bridge-test.yml
@@ -1,0 +1,40 @@
+---
+name: E2E Bridge Tests
+on: # yamllint disable-line rule:truthy
+  workflow_call:
+    outputs:
+      workflow_output:
+        description: E2E Bridge Tests output
+        value: ${{ jobs.e2e_bridge.outputs.e2e_bridge_output_failure }} 
+
+jobs:
+  e2e_bridge:
+    name: Run E2E Bridge Tests
+    runs-on: ubuntu-latest
+    env:
+      E2E_TESTS: true
+      E2E_LOGS: true
+      CI_VERBOSE: true
+    outputs:
+      e2e_bridge_output_failure: ${{ steps.run_e2e_bridge_failure.outputs.test_output }}
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4.1.1
+      - name: Install Go
+        uses: actions/setup-go@v5.0.0
+        with:
+          go-version: ${{ vars.GOLANG_VERSION }}
+          check-latest: true
+      - name: Run tests
+        run: make test-e2e-bridge
+      - name: Run tests failed
+        if: failure()
+        id: run_e2e_bridge_failure
+        run: echo "test_output=false" >> $GITHUB_OUTPUT
+      - name: Archive test logs
+        if: always()
+        uses: actions/upload-artifact@v4.3.0
+        with:
+          name: e2e-bridge-logs
+          path: e2e-logs-*/
+          retention-days: 30
